+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/PipelinesApiAsyncBackendJobExecutionActor.scala
@@ -26,7 +26,6 @@ import cromwell.backend.standard.{StandardAdHocValue, StandardAsyncExecutionActo
 import cromwell.core._
 import cromwell.core.path.{DefaultPathBuilder, Path}
 import cromwell.core.retry.SimpleExponentialBackoff
-import cromwell.filesystems.demo.dos.DemoDosPath
 import cromwell.filesystems.drs.DrsPath
 import cromwell.filesystems.gcs.GcsPath
 import cromwell.filesystems.gcs.batch.GcsBatchCommandBuilder
@@ -166,8 +165,7 @@ class PipelinesApiAsyncBackendJobExecutionActor(override val standardParams: Sta
   protected def relativeLocalizationPath(file: WomFile): WomFile = {
     file.mapFile(value =>
       getPath(value) match {
-        case Success(demoDosPath: DemoDosPath) =>
-          demoDosResolver.getContainerRelativePath(demoDosPath)
+        case Success(drsPath: DrsPath) => drsResolver.getContainerRelativePath(drsPath)
         case Success(path) => path.pathWithoutScheme
         case _ => value
       }
@@ -177,8 +175,7 @@ class PipelinesApiAsyncBackendJobExecutionActor(override val standardParams: Sta
   protected def fileName(file: WomFile): WomFile = {
     file.mapFile(value =>
       getPath(value) match {
-        case Success(demoDosPath: DemoDosPath) =>
-          DefaultPathBuilder.get(demoDosResolver.getContainerRelativePath(demoDosPath)).name
+        case Success(drsPath: DrsPath) => DefaultPathBuilder.get(drsResolver.getContainerRelativePath(drsPath)).name
         case Success(path) => path.name
         case _ => value
       }
@@ -693,10 +690,8 @@ class PipelinesApiAsyncBackendJobExecutionActor(override val standardParams: Sta
         case (Success(path @ ( _: GcsPath | _: HttpPath)), _) =>
           workingDisk.mountPoint.resolve(path.pathWithoutScheme).pathAsString
         case (Success(drsPath: DrsPath), _) =>
-          workingDisk.mountPoint.resolve(drsPath.pathWithoutScheme).pathAsString
-//        case (Success(demoDosPath: DemoDosPath), _) =>
-//          val filePath = demoDosResolver.getContainerRelativePath(demoDosPath)
-//          workingDisk.mountPoint.resolve(filePath).pathAsString
+          val filePath = drsResolver.getContainerRelativePath(drsPath)
+          workingDisk.mountPoint.resolve(filePath).pathAsString
         case (Success(sraPath: SraPath), _) =>
           workingDisk.mountPoint.resolve(s"sra-${sraPath.accession}/${sraPath.pathWithoutScheme}").pathAsString
         case _ => value
@@ -708,10 +703,9 @@ class PipelinesApiAsyncBackendJobExecutionActor(override val standardParams: Sta
     womFile.mapFile(value =>
       getPath(value) match {
         case Success(gcsPath: GcsPath) => workingDisk.mountPoint.resolve(gcsPath.pathWithoutScheme).pathAsString
-        case Success(drsPath: DrsPath) => workingDisk.mountPoint.resolve(drsPath.pathWithoutScheme).pathAsString
-//        case Success(demoDosPath: DemoDosPath) =>
-//          val filePath = demoDosResolver.getContainerRelativePath(demoDosPath)
-//          workingDisk.mountPoint.resolve(filePath).pathAsString
+        case Success(drsPath: DrsPath) =>
+          val filePath = drsResolver.getContainerRelativePath(drsPath)
+          workingDisk.mountPoint.resolve(filePath).pathAsString
         case _ => value
       }
     )
