+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/PipelinesApiJobCachingActorHelper.scala
@@ -6,7 +6,7 @@ import cromwell.backend.standard.StandardCachingActorHelper
 import cromwell.core.labels.Labels
 import cromwell.core.logging.JobLogging
 import cromwell.core.path.Path
-import cromwell.filesystems.demo.dos.DemoDosResolver
+import cromwell.filesystems.drs.DrsResolver
 import cromwell.services.metadata.CallMetadataKeys
 
 import scala.language.postfixOps
@@ -26,9 +26,7 @@ trait PipelinesApiJobCachingActorHelper extends StandardCachingActorHelper {
 
   lazy val workingDisk: PipelinesApiAttachedDisk = runtimeAttributes.disks.find(_.name == PipelinesApiWorkingDisk.Name).get
 
-  lazy val demoDosResolver: DemoDosResolver = DemoDosResolver(
-    initializationData.papiConfiguration.configurationDescriptor.globalConfig
-  )
+  lazy val drsResolver: DrsResolver = new DrsResolver()
 
   lazy val callRootPath: Path = pipelinesApiCallPaths.callExecutionRoot
   lazy val returnCodeFilename: String = pipelinesApiCallPaths.returnCodeFilename
